Extract payPaymentIntent to own file to satisfy bundle size limit
checkout.js was 2.72KB, exceeding the 2.5KB bundlemon limit.
Move payPaymentIntent (with its sponsored-fallback logic) to
payPaymentIntent.ts — checkout.ts re-exports it for API compat.
